+++ lib/htm/errors.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-# HTM (Hierarchical Temporary Memory) error classes
+# HTM (Hierarchical Temporal Memory) error classes
 #
 # All HTM errors inherit from HTM::Error, allowing you to catch
 # all HTM-related errors with a single rescue clause.